수정함:        211_check.sh 	수정함:        211_config.sh

+++ scripts/211_check.sh
@@ -72,12 +72,14 @@ _cpu_temp="0"
 FUNC_SEND_ROUTER_STS(){
     source "$_CUR_STS_FILE"
 
-    FUNC_MAKE_CURRENT_STATE_STR "${1}"
+	if [ "${_TELEGRAM_USE_FLAG}" = "1" ] ; then
+		FUNC_MAKE_CURRENT_STATE_STR "${1}"
 
-    #----------------------------------
-    FUNC_SEND_TELEGRAM_BOT "${_TELE_TOKEN}" "${_TELE_CHAT_ID}" "${_CURRENT_STATE_STR}"
+		#----------------------------------
+		FUNC_SEND_TELEGRAM_BOT "${_TELE_TOKEN}" "${_TELE_CHAT_ID}" "${_CURRENT_STATE_STR}"
 
-    _chk_ram_cnt_cur=0
+		_chk_ram_cnt_cur=0
+	fi
 }
 #######################################################################
 
@@ -115,7 +117,9 @@ do
         _tmp="$(nvram get territory_code)"
         if [ ${#_tmp} -gt 0 ] ; then
             if [ "${_tmp}" != "${_TERRITORY_CODE}" ] ; then
-                FUNC_SEND_TELEGRAM_BOT "${_TELE_TOKEN}" "${_TELE_CHAT_ID}" "territory_code 변경 ${_tmp} -> ${_TERRITORY_CODE}"
+				if [ "${_TELEGRAM_USE_FLAG}" = "1" ] ; then
+                	FUNC_SEND_TELEGRAM_BOT "${_TELE_TOKEN}" "${_TELE_CHAT_ID}" "territory_code 변경 ${_tmp} -> ${_TERRITORY_CODE}"
+				fi
                 nvram set territory_code="${_TERRITORY_CODE}"
                 nvram commit
             fi
@@ -130,7 +134,9 @@ do
         _tmp="$(nvram get location_code)"
         if [ ${#_tmp} -gt 0 ] ; then
             if [ "${_tmp}" != "${_LOCATION_CODE}" ] ; then
-                FUNC_SEND_TELEGRAM_BOT "${_TELE_TOKEN}" "${_TELE_CHAT_ID}" "location_code 변경 ${_tmp} -> ${_LOCATION_CODE}"
+				if [ "${_TELEGRAM_USE_FLAG}" = "1" ] ; then
+                	FUNC_SEND_TELEGRAM_BOT "${_TELE_TOKEN}" "${_TELE_CHAT_ID}" "location_code 변경 ${_tmp} -> ${_LOCATION_CODE}"
+				fi
                 nvram set location_code="${_LOCATION_CODE}"
                 nvram commit
             fi
@@ -141,15 +147,17 @@ do
     #-----------------------------------------------------------
     # 중국 uuplugin 관련 process 를 사용하지 않도록 kill
     #-----------------------------------------------------------
-    _tmp=$(pidof uuplugin_monitor.sh)
-    if [ "${_tmp}" ] ; then
-        kill -9 "${_tmp}"
-    fi
-    _tmp=$(pidof uuplugin)
-    if [ "${_tmp}" ] ; then
-        kill -9 "${_tmp}"
-    fi
-    sleep 1
+    if [ "${_KILL_UUPLUGIN}" = "1" ] ; then
+		_tmp=$(pidof uuplugin_monitor.sh)
+		if [ "${_tmp}" ] ; then
+			kill -9 "${_tmp}"
+		fi
+		_tmp=$(pidof uuplugin)
+		if [ "${_tmp}" ] ; then
+			kill -9 "${_tmp}"
+		fi
+		sleep 1
+	fi
 
     #-----------------------------------------------------------
     # openvpn 이 실행된 경우 내부망 접속을 위한 추가
@@ -164,7 +172,9 @@ do
             if [ ! "${_tmp}" ] ; then
                 #iptables -t nat -A POSTROUTING -s 192.168.10.0/24 -o br0 -j MASQUERADE
                 iptables -t nat -A POSTROUTING -s "${_vpn_ip}"/24 -o br0 -j MASQUERADE
-                FUNC_SEND_TELEGRAM_BOT "${_TELE_TOKEN}" "${_TELE_CHAT_ID}" "OPENVPN ${_vpn_ip} POSTROUTING ADDED"
+    			if [ "${_TELEGRAM_USE_FLAG}" = "1" ] ; then
+                	FUNC_SEND_TELEGRAM_BOT "${_TELE_TOKEN}" "${_TELE_CHAT_ID}" "OPENVPN ${_vpn_ip} POSTROUTING ADDED"
+				fi
             fi
         fi
     fi
@@ -173,93 +183,95 @@ do
     #-----------------------------------------------------------
     # ssh 접속을 위한 dropbear 이 죽는 경우 재실행 하도록 함
     #-----------------------------------------------------------
-    if [ ! "$(pidof dropbear)" ] ; then
-        _tmp="$(nvram get lan_ipaddr)"
-        if [ "${_tmp}" ] ; then
-            dropbear -p "${_tmp}":22 -a &
-            FUNC_SEND_TELEGRAM_BOT "${_TELE_TOKEN}" "${_TELE_CHAT_ID}" "dropbear ${_tmp} START"
-        fi
-    fi
-    sleep 1
+    #if [ ! "$(pidof dropbear)" ] ; then
+    #    _tmp="$(nvram get lan_ipaddr)"
+    #    if [ "${_tmp}" ] ; then
+    #        dropbear -p "${_tmp}":22 -a &
+    #		if [ "${_TELEGRAM_USE_FLAG}" = "1" ] ; then
+    #       		FUNC_SEND_TELEGRAM_BOT "${_TELE_TOKEN}" "${_TELE_CHAT_ID}" "dropbear ${_tmp} START"
+	#		fi
+    #    fi
+    #fi
+    #sleep 1
 
 
 
     #==============================================================================
     # 정각이 되거나 값이 변경되면 텔레그램으로 상태를 전송한다.
     #==============================================================================
-    if [ "${_TELEGRAM_USE_FLAG}" != "1" ] ; then
-        continue
-    fi
-    #------------------------------------------------------------------------------
-    # 매일 0900 시에 공유기 상태를 텔레그램으로 전송한다.
-    #------------------------------------------------------------------------------
-    _tmp="$(date +%H%M)"
-    if [ "$_tmp" = "0900" ] || [ "$_tmp" = "0901" ] ; then
-        if [ $_my_sts_sent_flag -eq 0 ] ; then
-            _my_sts_sent_flag=1
-            FUNC_SEND_ROUTER_STS "오전 9시"
-            continue
-        fi
-    else
-        _my_sts_sent_flag=0
-    fi
-    sleep 1
+    if [ "${_TELEGRAM_USE_FLAG}" = "1" ] ; then
 
-    #------------------------------------------------------------------------------
-    # CPU 온도 변화 체크 -> 공유기 상태를 텔레그램으로 전송한다.
-    #------------------------------------------------------------------------------
-    _tmp=$(awk '{printf("%d\n", $1/1000 + 0.5)}' < /sys/class/thermal/thermal_zone0/temp)
-    if [ ${#_tmp} -gt 0 ] ; then
-        _tmp_max=$((_cpu_temp + _DIFF_TEMP))
-        _tmp_min=$((_cpu_temp - _DIFF_TEMP))
-        if [ "${_tmp}" -ge "${_tmp_max}" ] || [ "${_tmp}" -le "${_tmp_min}" ] ; then
-            FUNC_SEND_ROUTER_STS "Cpu Temp. ${_cpu_temp} -> ${_tmp}"
-            _cpu_temp="${_tmp}"
-            continue
-        fi
-    fi
-    sleep 1
+		#------------------------------------------------------------------------------
+		# 매일 0900 시에 공유기 상태를 텔레그램으로 전송한다.
+		#------------------------------------------------------------------------------
+		_tmp="$(date +%H%M)"
+		if [ "$_tmp" = "0900" ] || [ "$_tmp" = "0901" ] ; then
+			if [ $_my_sts_sent_flag -eq 0 ] ; then
+				_my_sts_sent_flag=1
+				FUNC_SEND_ROUTER_STS "오전 9시"
+				continue
+			fi
+		else
+			_my_sts_sent_flag=0
+		fi
+		sleep 1
 
-    #------------------------------------------------------------------------------
-    # RAM FREE 용량 체크 -> 공유기 상태를 텔레그램으로 전송한다.
-    #------------------------------------------------------------------------------
-    _tmp=$(free | grep Mem | awk '{printf("%d", $4/1000 + 0.5)}')
-    if [ ${#_tmp} -gt 0 ] ; then
-        _tmp_max=$((_ram + _DIFF_RAM))
-        _tmp_min=$((_ram - _DIFF_RAM))
-        if [ "${_tmp}" -ge "${_tmp_max}" ] || [ "${_tmp}" -le "${_tmp_min}" ] ; then
-            _chk_ram_cnt_cur=$((_chk_ram_cnt_cur + 1))
-            if [ "${_chk_ram_cnt_cur}" -ge "${_CHK_RAM_CNT_MAX}" ] ; then
-                FUNC_SEND_ROUTER_STS "Mem. Free ${_ram} -> ${_tmp}"
-                _ram="${_tmp}"
-                continue
-            fi
-        else
-            _chk_ram_cnt_cur=0
-        fi
-    fi
-    sleep 1
+		#------------------------------------------------------------------------------
+		# CPU 온도 변화 체크 -> 공유기 상태를 텔레그램으로 전송한다.
+		#------------------------------------------------------------------------------
+		_tmp=$(awk '{printf("%d\n", $1/1000 + 0.5)}' < /sys/class/thermal/thermal_zone0/temp)
+		if [ ${#_tmp} -gt 0 ] ; then
+			_tmp_max=$((_cpu_temp + _DIFF_TEMP))
+			_tmp_min=$((_cpu_temp - _DIFF_TEMP))
+			if [ "${_tmp}" -ge "${_tmp_max}" ] || [ "${_tmp}" -le "${_tmp_min}" ] ; then
+				FUNC_SEND_ROUTER_STS "Cpu Temp. ${_cpu_temp} -> ${_tmp}"
+				_cpu_temp="${_tmp}"
+				continue
+			fi
+		fi
+		sleep 1
 
-    #-----------------------------------------------------------
-    # WAN_IP 외부에서 공유기 접속을 위한 IP를 구해서 텔레그램으로 전송
-    #-----------------------------------------------------------
-    #분을 10으로 나눈 값이 0이면...-> 10분단위로 체크하기 위함
-    if [ "$(expr "$(date +%M)" % 10)" = "0" ] ; then
-        if [ $_wan_ip_chk_flag -eq 0 ] ; then
-            _tmp=$(wget -O - -q http://checkip.dynu.com | awk -F": " '{print $2}')
-            if [ ${#_tmp} -gt 0 ] ; then
-                if [ "${_tmp}" != "${_wan_ip}" ] ; then
-                    FUNC_SEND_ROUTER_STS "WAN IP"
-                    _wan_ip="${_tmp}"
-                    continue
-                fi
-                _wan_ip_chk_flag=1
-            fi
-        fi
-    else
-        _wan_ip_chk_flag=0
+		#------------------------------------------------------------------------------
+		# RAM FREE 용량 체크 -> 공유기 상태를 텔레그램으로 전송한다.
+		#------------------------------------------------------------------------------
+		_tmp=$(free | grep Mem | awk '{printf("%d", $4/1000 + 0.5)}')
+		if [ ${#_tmp} -gt 0 ] ; then
+			_tmp_max=$((_ram + _DIFF_RAM))
+			_tmp_min=$((_ram - _DIFF_RAM))
+			if [ "${_tmp}" -ge "${_tmp_max}" ] || [ "${_tmp}" -le "${_tmp_min}" ] ; then
+				_chk_ram_cnt_cur=$((_chk_ram_cnt_cur + 1))
+				if [ "${_chk_ram_cnt_cur}" -ge "${_CHK_RAM_CNT_MAX}" ] ; then
+					FUNC_SEND_ROUTER_STS "Mem. Free ${_ram} -> ${_tmp}"
+					_ram="${_tmp}"
+					continue
+				fi
+			else
+				_chk_ram_cnt_cur=0
+			fi
+		fi
+		sleep 1
+
+		#-----------------------------------------------------------
+		# WAN_IP 외부에서 공유기 접속을 위한 IP를 구해서 텔레그램으로 전송
+		#-----------------------------------------------------------
+		#분을 10으로 나눈 값이 0이면...-> 10분단위로 체크하기 위함
+		if [ "$(expr "$(date +%M)" % 10)" = "0" ] ; then
+			if [ $_wan_ip_chk_flag -eq 0 ] ; then
+				_tmp=$(wget -O - -q http://checkip.dynu.com | awk -F": " '{print $2}')
+				if [ ${#_tmp} -gt 0 ] ; then
+					if [ "${_tmp}" != "${_wan_ip}" ] ; then
+						FUNC_SEND_ROUTER_STS "WAN IP"
+						_wan_ip="${_tmp}"
+						continue
+					fi
+					_wan_ip_chk_flag=1
+				fi
+			fi
+		else
+			_wan_ip_chk_flag=0
+		fi
+		sleep 1
     fi
-    sleep 1
 
 
     #-----------------------------------------------------------
@@ -277,6 +289,7 @@ do
         fi
     fi
     sleep 1
+
 
 
 done
